refactor:
- duration option
- Carousel.js -> index.js
- CSS_PATH, FETCH_PATH
- Transition on,off method
